check permissions export
