+++ app/functions/src/chtype.ts
@@ -1,7 +1,7 @@
 import * as functions from 'firebase-functions'
 import * as admin from 'firebase-admin';
-import { User } from './models/user.model';
 import { chType } from './models/chType.model';
+import { checkPermissions } from './permissions'
 
 export const addNewChtype = functions.https.onCall(async (chtype: chType, context) => {
   await checkPermissions(context);
@@ -139,27 +139,3 @@ export const deleteChtype = functions.https.onCall(async (chtype : chType, conte
     .doc(chtype.id)
     .delete();
 })
-
-
-async function checkPermissions(context: functions.https.CallableContext) {
-  // if user is not authenticated, he/she must authenticate
-  if (!context.auth) {
-    throw new functions.https.HttpsError(
-      'unauthenticated',
-      'must log in as admin'
-    )
-  }
-
-  // only admin can edit cultural heritage type
-  const snapshot = await admin.firestore()
-    .collection('users')
-    .doc(context.auth.uid)
-    .get();
-  const user: User = snapshot.data() as User;
-  if (user.role !== 'admin') {
-    throw new functions.https.HttpsError(
-      'permission-denied',
-      'not logged in as admin'
-    )
-  }
-}
