+++ app/functions/src/cultural-heritage.ts
@@ -1,8 +1,7 @@
 import * as functions from 'firebase-functions'
 import * as admin from 'firebase-admin';
 import { CulturalHeritage } from './models/culturalHeritage.model';
-import { User } from './models/user.model';
-
+import { checkPermissions } from './permissions'
 
 export const addNewCulturalHeritage = functions.https.onCall(async (ch: CulturalHeritage, context) => {
   await checkPermissions(context);
@@ -46,25 +45,3 @@ export const deleteCulturalHeritage = functions.https.onCall(async (ch: Cultural
     .delete();
 })
 
-async function checkPermissions(context: functions.https.CallableContext) {
-  // if user is not authenticated, he/she must authenticate
-  if (!context.auth) {
-    throw new functions.https.HttpsError(
-      'unauthenticated',
-      'must log in as admin'
-    )
-  }
-
-  // only admin can edit cultural heritage type
-  const snapshot = await admin.firestore()
-    .collection('users')
-    .doc(context.auth.uid)
-    .get();
-  const user: User = snapshot.data() as User;
-  if (user.role !== 'admin') {
-    throw new functions.https.HttpsError(
-      'permission-denied',
-      'not logged in as admin'
-    )
-  }
-}
